refactor to support customizable featurizers; add qualx_config
