+++ user_tools/tests/spark_rapids_tools_ut/qualx/test_config.py
@@ -15,6 +15,7 @@
 """Test qualx_config module"""
 import pytest  # pylint: disable=import-error
 from spark_rapids_tools.tools.qualx.config import (
+    get_config,
     get_cache_dir,
     get_label,
 )
@@ -30,18 +31,22 @@ class TestConfig(SparkRapidsToolsUT):
 
         # Test without environment variable (should use default)
         monkeypatch.delenv('QUALX_CACHE_DIR')
+        get_config(reload=True)  # reload config
         assert get_cache_dir() == 'qualx_cache'
 
     def test_get_label(self, monkeypatch):
         # Test with duration_sum
         monkeypatch.setenv('QUALX_LABEL', 'duration_sum')
+        get_config(reload=True)  # reload config
         assert get_label() == 'duration_sum'
 
         # Test with unsupported label
-        with pytest.raises(AssertionError):
+        with pytest.raises(ValueError):
             monkeypatch.setenv('QUALX_LABEL', 'duration')
+            get_config(reload=True)  # reload config
             get_label()
 
         # Test without environment variable (should use default)
         monkeypatch.delenv('QUALX_LABEL')
+        get_config(reload=True)  # reload config
         assert get_label() == 'Duration'
